backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
xgboost==2.1.0
shap==0.45.0
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.0
pdfplumber==0.11.0
python-multipart==0.0.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy>=2.0.0
aiosqlite>=0.20.0
